Actualizacion Linterna

+++ TP-FPS-LeonD/Assets/Scripts/DañoBala.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public class DañoBala : MonoBehaviour
 {
-    public int CantidadDaño = 25;
+    public float CantidadDaño = 25f;
 }
